pyproject.toml
[project]
name = "willgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.2",
    "FlagEmbedding>=1.2.0",
    "tqdm>=4.67.1",
    "black>=25.9.0",
]
